third_party/WebKit/LayoutTests/fast/dom/idl-dictionary-unittest.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<script>
description('IDL dictionary unittest');

if (window.internals && internals.dictionaryTest) {
    var dictionaryTest = window.internals.dictionaryTest();

    debug('Test for setting an empty dictionary');
    dictionaryTest.set({});
    dict = dictionaryTest.get();
    shouldBeUndefined('dict.longMember');
    shouldBe('dict.longMemberWithDefault', '42');
    shouldBeUndefined('dict.longOrNullMember');
    shouldBeNull('dict.longOrNullMemberWithDefault');
    shouldBeUndefined('dict.booleanMember');
    shouldBeUndefined('dict.doubleMember');
    shouldBeUndefined('dict.stringMember');
    shouldBeEqualToString('dict.stringMemberWithDefault', 'defaultStringValue');
    shouldBeUndefined('dict.stringSequenceMember');
    shouldBeUndefined('dict.stringSequenceOrNullMember');
    shouldBeUndefined('dict.elementMember');
    shouldBeUndefined('dict.elementOrNullMember');
    shouldBeUndefined('dict.enumMember');
    shouldBeEqualToString('dict.enumMemberWithDefault', 'foo');
    shouldBeUndefined('dict.enumOrNullMember');
    debug('');

    debug('Test for setting undefined');
    dictionaryTest.set(undefined);
    dict = dictionaryTest.get();
    shouldBeUndefined('dict.longMember');
    shouldBe('dict.longMemberWithDefault', '42');
    shouldBeUndefined('dict.longOrNullMember');
    shouldBeNull('dict.longOrNullMemberWithDefault');
    shouldBeUndefined('dict.booleanMember');
    shouldBeUndefined('dict.doubleMember');
    shouldBeUndefined('dict.stringMember');
    shouldBeEqualToString('dict.stringMemberWithDefault', 'defaultStringValue');
    shouldBeUndefined('dict.stringSequenceMember');
    shouldBeUndefined('dict.stringSequenceOrNullMember');
    shouldBeUndefined('dict.elementMember');
    shouldBeUndefined('dict.elementOrNullMember');
    debug('');

    var element1 = document.createElement('div');

    debug('Test for setting valid values');
    dictionaryTest.set({
        longMember: 1,
        longMemberWithDefault: 2,
        longOrNullMember: 3,
        longOrNullMemberWithDefault: 4,
        stringMember: 'modifiedString1',
        stringMemberWithDefault: 'modifiedString2',
        booleanMember: true,
        doubleMember: 3.14,
        stringSequenceMember: ['foo', 'bar', 'baz'],
        stringSequenceOrNullMember: [],
        elementMember: element1,
        elementOrNullMember: null,
        enumMember: 'foo',
        enumMemberWithDefault: 'bar',
        enumOrNullMember: 'baz',
    });
    dict = dictionaryTest.get();
    shouldBe('dict.longMember', '1');
    shouldBe('dict.longMemberWithDefault', '2');
    shouldBe('dict.longOrNullMember', '3');
    shouldBe('dict.longOrNullMemberWithDefault', '4');
    shouldBeEqualToString('dict.stringMember', 'modifiedString1');
    shouldBeEqualToString('dict.stringMemberWithDefault', 'modifiedString2');
    shouldBeTrue('dict.booleanMember');
    shouldBe('dict.doubleMember', '3.14');
    shouldBe('dict.stringSequenceMember', '["foo", "bar", "baz"]');
    shouldBe('dict.stringSequenceOrNullMember', '[]');
    shouldBe('dict.elementMember', 'element1');
    // This will be undefined as the spec says
    shouldBeUndefined('dict.elementOrNullMember');
    shouldBeEqualToString('dict.enumMember', 'foo');
    shouldBeEqualToString('dict.enumMemberWithDefault', 'bar');
    shouldBeEqualToString('dict.enumOrNullMember', 'baz');
    debug('');

    debug('Test for explicit undefined or null, and missing members');
    dictionaryTest.set({
        longMember: undefined,
        longMemberWithDefault: null,
        longOrNullMember: undefined,
        longOrNullMemberWithDefault: null,
    });
    dict = dictionaryTest.get();
    shouldBeUndefined('dict.longMember');
    shouldBe('dict.longMemberWithDefault', '42');
    shouldBeUndefined('dict.longOrNullMember');
    // Nullable and its default value is null
    shouldBeNull('dict.longOrNullMemberWithDefault');
    debug('');

    debug('Test for setting invalid member');
    dictionaryTest.set({invalidMember: 'shouldNotBeSet'});
    dict = dictionaryTest.get();
    shouldBeUndefined('dict.invalidMember');
    debug('');

    debug('Test for setting invalid enum value');
    shouldThrow("dictionaryTest.set({enumMember: 'invalid'})");
    debug('');
}
</script>
